Restructure test target
- Split tests by domain
- Group test data set files
